Added a warning to lab 4.2 for when a non-accepted string causes the machine complete a computation successfully, run forever or go off the tape.
Lab 4 calls for inputs that don't give an answer to "terminate without completing a computation". I've double checked, and running forever, running off the left edge, or  ending in an accepting state with the head parked are not acceptable ways of doing this. The verifier should now show a warning for all these cases on inputs where the output doesn't matter.

Also fixed bug where the output displayed that the machine had halted if it ran for >= maxsteps.